# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorResponse:
    properties:
      code:
        type: integer
      detail:
        type: string
      error:
        type: string
    type: object
  model.LocationUpdate:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      ref_id:
        type: string
    required:
    - latitude
    - longitude
    - ref_id
    type: object
  model.NearbyEntity:
    properties:
      distance_km:
        type: number
      latitude:
        type: number
      longitude:
        type: number
      ref_id:
        type: string
    type: object
  model.TrackedEntity:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      ref_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8083
info:
  contact: {}
  description: |-
    Real-time GPS tracking service with Redis geo indexing.
    Generic entity tracking — accepts any ref_id (driver, rider, courier, etc).
  title: StudEx Location Service API
  version: "1.0"
paths:
  /location:
    get:
      description: Get current positions of all tracked entities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TrackedEntity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get all tracked locations
      tags:
      - location
    post:
      consumes:
      - application/json
      description: Update an entity's current GPS position. Called every 2-5 seconds
        by clients.
      parameters:
      - description: Location data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LocationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Update entity location
      tags:
      - location
  /location/{ref_id}:
    delete:
      description: Remove an entity's location from the tracking system
      parameters:
      - description: Reference ID
        in: path
        name: ref_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove entity location
      tags:
      - location
    get:
      description: Get an entity's current GPS position
      parameters:
      - description: Reference ID
        in: path
        name: ref_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TrackedEntity'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Get entity location
      tags:
      - location
  /location/nearby:
    get:
      description: Find all entities within a given radius of a point
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lng
        required: true
        type: number
      - default: 5
        description: Radius in km
        in: query
        name: radius
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.NearbyEntity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Find nearby entities
      tags:
      - location
schemes:
- http
- https
swagger: "2.0"